+++ resources/views/accounting/transactions/index.blade.php
@@ -240,6 +240,7 @@
                             .show();
                     }
                     @include('accounting.transactions.controlled')
+                    refreshFsLightbox()
                 });
             });
         });