+++ Changelog.md
@@ -9,6 +9,7 @@
 * Removed unnecesary coupon type details view
 * Only show webling booking button if configured
 * Removed google books PHP library, replaced with Javascript solution
+* Replaced lity lightbox with fsightbox which has a gallery feature
 
 ## 2.29.0